# openapi/backend.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/cases":{"get":{"summary":"Get All Cases","operationId":"get_all_cases_cases_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Case"},"type":"array","title":"Response Get All Cases Cases Get"}}}}}},"post":{"summary":"Post Cases","operationId":"post_cases_cases_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Case"}}}}}}},"/cases/{case_id}":{"get":{"summary":"Get Case By Id","operationId":"get_case_by_id_cases__case_id__get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Case"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Case":{"properties":{"case_id":{"type":"integer","title":"Case Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"status":{"$ref":"#/components/schemas/Status"},"procedure_name":{"type":"string","title":"Procedure Name"},"cpt_codes":{"items":{"type":"string"},"type":"array","title":"Cpt Codes"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"},"is_met":{"type":"boolean","title":"Is Met"},"is_complete":{"type":"boolean","title":"Is Complete"},"steps":{"items":{"$ref":"#/components/schemas/Step"},"type":"array","title":"Steps"}},"type":"object","required":["case_id","created_at","status","procedure_name","cpt_codes","summary","is_met","is_complete","steps"],"title":"Case"},"Evidence":{"properties":{"content":{"type":"string","title":"Content"},"page_number":{"type":"integer","title":"Page Number"},"pdf_name":{"type":"string","title":"Pdf Name"},"event_datetime":{"type":"string","format":"date-time","title":"Event Datetime"}},"type":"object","required":["content","page_number","pdf_name","event_datetime"],"title":"Evidence"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Option":{"properties":{"key":{"type":"string","title":"Key"},"text":{"type":"string","title":"Text"},"selected":{"type":"boolean","title":"Selected"}},"type":"object","required":["key","text","selected"],"title":"Option"},"Status":{"type":"string","enum":["unknown","submitted","processing","completed"],"title":"Status"},"Step":{"properties":{"key":{"type":"string","title":"Key"},"question":{"type":"string","title":"Question"},"options":{"items":{"$ref":"#/components/schemas/Option"},"type":"array","title":"Options"},"reasoning":{"type":"string","title":"Reasoning"},"decision":{"type":"string","title":"Decision"},"next_step":{"type":"string","title":"Next Step"},"is_met":{"type":"boolean","title":"Is Met"},"is_final":{"type":"boolean","title":"Is Final"},"evidence":{"items":{"$ref":"#/components/schemas/Evidence"},"type":"array","title":"Evidence"}},"type":"object","required":["key","question","options","reasoning","decision","next_step","is_met","is_final","evidence"],"title":"Step"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}